Add files via upload
Since the data set is huge, it will be difficult for you to download it and work on it offline. Hence, we have downloaded the raw data set and shared it with you at this Google Drive link. Please do not create any files in this drive folder. You are advised to use Google Colab to write your Python code. Create a separate folder in your Google Drive to save any code/CSV files (processed data) and the colab notebooks that you create during this project.

https://drive.google.com/drive/folders/1LMU-jl0DgV4qcQKuO9P-R9Asd1_pexmV